+++ src/main/java/ca/corefacility/bioinformatics/irida/model/project/Project.java
@@ -4,7 +4,6 @@ import ca.corefacility.bioinformatics.irida.model.IridaResourceSupport;
 import ca.corefacility.bioinformatics.irida.model.MutableIridaThing;
 import ca.corefacility.bioinformatics.irida.model.NcbiExportSubmission;
 import ca.corefacility.bioinformatics.irida.model.event.ProjectEvent;
-import ca.corefacility.bioinformatics.irida.model.irida.IridaProject;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectMetadataTemplateJoin;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectSampleJoin;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectUserJoin;
@@ -15,9 +14,11 @@ import ca.corefacility.bioinformatics.irida.model.user.group.UserGroupProjectJoi
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmission;
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.AnalysisSubmissionTemplate;
 import ca.corefacility.bioinformatics.irida.model.workflow.submission.ProjectAnalysisSubmissionJoin;
+import ca.corefacility.bioinformatics.irida.validators.annotations.ValidProjectName;
 import org.codehaus.jackson.annotate.JsonIgnore;
 import org.hibernate.envers.Audited;
 import org.hibernate.envers.NotAudited;
+import org.hibernate.validator.constraints.URL;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
@@ -25,6 +26,7 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import javax.persistence.*;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
@@ -38,12 +40,15 @@ import java.util.Objects;
 @Audited
 @EntityListeners(AuditingEntityListener.class)
 public class Project extends IridaResourceSupport
-		implements MutableIridaThing, IridaProject, Comparable<Project>, RemoteSynchronizable {
+		implements MutableIridaThing, Comparable<Project>, RemoteSynchronizable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
+	@NotNull(message = "{project.name.notnull}")
+	@Size(min = 5, message = "{project.name.size}")
+	@ValidProjectName
 	private String name;
 
 	@CreatedDate
@@ -60,6 +65,7 @@ public class Project extends IridaResourceSupport
 	@Lob
 	private String projectDescription;
 
+	@URL(message = "{project.remoteURL.url}")
 	private String remoteURL;
 
 	private String organism;