+++ src/main/java/ca/corefacility/bioinformatics/irida/model/sample/Sample.java
@@ -3,18 +3,21 @@ package ca.corefacility.bioinformatics.irida.model.sample;
 import ca.corefacility.bioinformatics.irida.model.IridaResourceSupport;
 import ca.corefacility.bioinformatics.irida.model.MutableIridaThing;
 import ca.corefacility.bioinformatics.irida.model.event.SampleAddedProjectEvent;
-import ca.corefacility.bioinformatics.irida.model.irida.IridaSample;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.ProjectSampleJoin;
 import ca.corefacility.bioinformatics.irida.model.joins.impl.SampleGenomeAssemblyJoin;
 import ca.corefacility.bioinformatics.irida.model.remote.RemoteStatus;
 import ca.corefacility.bioinformatics.irida.model.remote.RemoteSynchronizable;
 import ca.corefacility.bioinformatics.irida.model.sample.metadata.MetadataEntry;
+import ca.corefacility.bioinformatics.irida.validators.annotations.Latitude;
+import ca.corefacility.bioinformatics.irida.validators.annotations.Longitude;
+import ca.corefacility.bioinformatics.irida.validators.annotations.ValidSampleName;
+import ca.corefacility.bioinformatics.irida.validators.groups.NCBISubmission;
+import ca.corefacility.bioinformatics.irida.validators.groups.NCBISubmissionOneOf;
 import ca.corefacility.bioinformatics.irida.web.controller.api.json.DateJson;
-
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-
 import org.hibernate.envers.Audited;
 import org.hibernate.envers.NotAudited;
 import org.springframework.data.annotation.CreatedDate;
@@ -23,6 +26,8 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+import javax.validation.constraints.Size;
 import java.util.*;
 import java.util.Map.Entry;
 
@@ -40,8 +45,9 @@ import java.util.Map.Entry;
 @Table(name = "sample")
 @Audited
 @EntityListeners(AuditingEntityListener.class)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class Sample extends IridaResourceSupport
-		implements MutableIridaThing, IridaSample, Comparable<Sample>, RemoteSynchronizable {
+		implements MutableIridaThing, Comparable<Sample>, RemoteSynchronizable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -57,6 +63,9 @@ public class Sample extends IridaResourceSupport
 	@Temporal(TemporalType.TIMESTAMP)
 	private Date modifiedDate;
 
+	@NotNull(message = "{sample.name.notnull}")
+	@Size(min = 3, message = "{sample.name.too.short}")
+	@ValidSampleName
 	private String sampleName;
 
 	@Lob
@@ -66,22 +75,30 @@ public class Sample extends IridaResourceSupport
 	 * The most descriptive organism name for this sample (to the species, if
 	 * relevant).
 	 */
+	@NotNull(message = "{sample.organism.notnull}", groups = NCBISubmission.class)
+	@Size(min = 3, message = "{sample.organism.too.short}")
 	private String organism;
 
 	/**
 	 * identification or description of the specific individual from which this
 	 * sample was obtained
 	 */
+	@NotNull(message = "{sample.isolate.notnull}", groups = { NCBISubmission.class, NCBISubmissionOneOf.class })
+	@Size(min = 3, message = "{sample.isolate.too.short}")
 	private String isolate;
 
 	/**
 	 * microbial or eukaryotic strain name
 	 */
+	@NotNull(message = "{sample.strain.name.notnull}", groups = { NCBISubmission.class, NCBISubmissionOneOf.class })
+	@Size(min = 3, message = "{sample.strain.name.too.short}")
 	private String strain;
 
 	/**
 	 * Name of the person who collected the sample.
 	 */
+	@NotNull(message = "{sample.collected.by.notnull}", groups = NCBISubmission.class)
+	@Size(min = 3, message = "{sample.collected.by.too.short}")
 	private String collectedBy;
 
 	/**
@@ -90,12 +107,16 @@ public class Sample extends IridaResourceSupport
 	@Temporal(TemporalType.DATE)
 	@JsonSerialize(as=java.sql.Date.class, using = DateJson.DateSerializer.class)
 	@JsonDeserialize(as=java.sql.Date.class, using = DateJson.DateDeserializer.class)
+	@NotNull(message = "{sample.collection.date.notnull}", groups = NCBISubmission.class)
 	private Date collectionDate;
 
 	/**
 	 * Geographical origin of the sample (country derived from
 	 * http://www.insdc.org/documents/country-qualifier-vocabulary).
 	 */
+	@NotNull(message = "{sample.geographic.location.name.notnull}", groups = NCBISubmission.class)
+	@Pattern(regexp = "\\w+(:\\w+(:\\w+)?)?", message = "{sample.geographic.location.name.pattern}")
+	@Size(min = 3, message = "{sample.geographic.location.name.too.short}")
 	private String geographicLocationName;
 
 	/**
@@ -103,14 +124,19 @@ public class Sample extends IridaResourceSupport
 	 * the biological sample from which the sample was derived.
 	 */
 	@Lob
+	@NotNull(message = "{sample.isolation.source.notnull}", groups = NCBISubmission.class)
 	private String isolationSource;
 
 	/**
 	 * lat_lon is marked as a *mandatory* attribute in NCBI BioSample, but in
 	 * practice many of the fields are shown as "missing".
 	 */
+	@NotNull(message = "{sample.latitude.notnull}", groups = NCBISubmission.class)
+	@Latitude
 	private String latitude;
 
+	@NotNull(message = "{sample.longitude.notnull}", groups = NCBISubmission.class)
+	@Longitude
 	private String longitude;
 
 	@OneToMany(fetch = FetchType.LAZY, cascade = CascadeType.REMOVE, mappedBy = "sample")