+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/linelist/LineListController.java
@@ -36,6 +36,8 @@ import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
 
+import com.google.common.base.Strings;
+
 /**
  * This controller is responsible for AJAX handling for the line list page, which displays sample metadata.
  */
@@ -254,8 +256,9 @@ public class LineListController {
 				MetadataTemplateField metadataTemplateField = metadataTemplateService.readMetadataFieldByKey(
 						field.getField());
 				if (metadataTemplateField == null) {
+					String type = Strings.isNullOrEmpty(field.getType()) ? "text" : field.getType();
 					metadataTemplateField = metadataTemplateService.saveMetadataField(
-							new MetadataTemplateField(field.getHeaderName(), field.getType()));
+							new MetadataTemplateField(field.getHeaderName(), type));
 				}
 				fields.add(metadataTemplateField);
 			}