+++ src/main/java/ca/corefacility/bioinformatics/irida/service/export/ExportUploadService.java
@@ -306,6 +306,7 @@ public class ExportUploadService {
 					} else {
 						filename = file.getSequenceFile().getId() + ".fastq";
 					}
+
 					uploadPath(client, filename, file.getSequenceFile().getFile());
 				}