+++ src/main/java/ca/corefacility/bioinformatics/irida/service/impl/ProjectServiceImpl.java
@@ -696,13 +696,15 @@ public class ProjectServiceImpl extends CRUDServiceImpl<Long, Project> implement
 
 			// get the projects for the samples
 			for (SampleSequencingObjectJoin s : samples) {
-				psjRepository.getProjectForSample(s.getSubject())
-						.forEach(p -> {
-							// p may be null if sample was removed from all projects
-							if (p != null) {
-								projects.add(p.getSubject());
-							}
-						});
+				if(s != null) {
+					psjRepository.getProjectForSample(s.getSubject())
+							.forEach(p -> {
+								// p may be null if sample was removed from all projects
+								if (p != null) {
+									projects.add(p.getSubject());
+								}
+							});
+				}
 			}
 		}